requirements.txt
opencv-python
numpy
scipy
python-telegram-bot
python-dotenv